[PetTracker] added exceptions and null checks

+++ src/main/java/shop/pet/track/service/PetServiceImpl.java
@@ -1,7 +1,6 @@
 package shop.pet.track.service;
 
 import jakarta.transaction.Transactional;
-import org.hibernate.tool.schema.spi.SqlScriptException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import shop.pet.track.dao.OwnerDAO;
@@ -12,8 +11,9 @@ import shop.pet.track.exception.NotFoundException;
 
 import java.time.LocalDate;
 import java.util.Set;
+
 @Service
-public class PetServiceImpl implements PetService{
+public class PetServiceImpl implements PetService {
     OwnerDAO ownerDAO;
     PetDAO petDAO;
 
@@ -26,52 +26,59 @@ public class PetServiceImpl implements PetService{
     @Override
     @Transactional
     public void addPetsToOwner(Integer id, Pet pet) {
-        try {
-            Owner owner = ownerDAO.find(id);
+
+        Owner owner = ownerDAO.find(id);
+        if (owner == null) {
+            throw new NotFoundException("owner not found id :" + id);
+        } else {
             LocalDate localDate = LocalDate.now();
             pet.setOwner(owner);
             pet.setDateCreated(localDate);
             petDAO.addPet(pet);
-        } catch (Exception ex) {
-            throw new NotFoundException("owner not found id :" + id);
         }
-
     }
 
     @Override
     @Transactional
     public void deletePet(Integer id) {
-        try {
-            Pet pet = petDAO.find(id);
-            pet.setOwner(null);
+        Pet pet = petDAO.find(id);
+        if (pet != null) {
             petDAO.delete(pet);
-        } catch (Exception ex) {
-            throw new NotFoundException("Pet not found" + id);
+        } else {
+            throw new NotFoundException("Pet not found id " + id);
         }
     }
 
     @Override
     @Transactional
-    public void updatePet(Integer id, String name, String breed) {
-        try {
+    public void updatePet(Integer id, Pet pet) {
+        Pet myPet = petDAO.find(id);
+        if (myPet != null) {
+            if (!pet.getBreed().isEmpty()) {
+                myPet.setBreed(pet.getBreed());
+            }
+            if (!pet.getName().isEmpty()) {
+                myPet.setName(pet.getName());
+            }
             LocalDate localDate = LocalDate.now();
-            Pet pet = petDAO.find(id);
-            pet.setDateModified(localDate);
-            pet.setBreed(breed);
-            pet.setName(name);
-            petDAO.update(pet);
-        } catch (Exception ex) {
-            throw new NotFoundException("Pet not found" + id);
+            myPet.setDateModified(localDate);
+            petDAO.update(myPet);
+        } else {
+            throw new NotFoundException("Pet not found id " + id);
         }
     }
 
     @Override
     public Set<Pet> getPetsByOwnerId(Integer id) {
-        try {
-            Set<Pet> pets  = petDAO.getPetsByOwnerId(id);
+        Owner owner = ownerDAO.find(id);
+        if (owner != null) {
+            Set<Pet> pets = petDAO.getPetsByOwnerId(id);
+            if (pets.isEmpty()) {
+                throw new NotFoundException("Owner not found" + id);
+            }
             return pets;
-        } catch (SqlScriptException ex) {
-            throw new NotFoundException("Owner not found" + id);
+        } else {
+            throw new NotFoundException("Owner not found id " + id);
         }
     }
 }
